requirements.txt
qq-botpy>=1.1.2
PyMySQL~=1.1.0
aiocron~=1.8
aiohttp~=3.8.4
feedparser~=6.0.10
PyYAML~=6.0
requests~=2.28.2
Pillow~=9.5.0
mcstatus~=11.0.0
redis~=4.5.4
PyMysqlPool~=1.0.9
DBUtils==1.3
pycryptodome
bs4~=0.0.1
beautifulsoup4~=4.12.2
mysql-connector-python~=8.1.0